Added IsValid_IP_Address project
Worked on this problem with Jon Cambell. Notable points are that
Split('.') returns a collection of strings from the input (I.e.,
"192.155.10.5".Split('.') returns a set of strings = {"192", "155",
"10", "5"}.

Convert.ToInt32 will fail when a character is given, so it's better to
use Int32.TryParse(string s, out int output).

A foreach loop is available because the collection of objects have
iterators.

C# is object-oriented, interpreted and contains garbage collection.